added routes

+++ migrations/20180424131358_employee.js
@@ -5,7 +5,7 @@ exports.up = function(knex, Promise) {
         table.string('first_name', 255);
         table.string('last_name', 255);
         table.string('email', 60);
-        table.integer('phone_number');
+        table.bigInteger('phone_number');
         table.string('google_token', 255);
         table.integer('security_level', 5);
         table.timestamps();
